requirements.txt
Django==1.11.17,<3.0
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
Markdown==3.1.1
psycopg2>=2.7,<3.0
pytz==2019.3


# heroku configuration
gunicorn==19.9.0
whitenoise==4.1.4
dj-database-url==0.5.0


# talking to React app
django-cors-headers==3.0.0